[JSON] make API public but experimental
