+++ src/core/lib/security/credentials/jwt/json_token.cc
@@ -170,7 +170,7 @@ static char* encoded_jwt_header(const char* key_id, const char* algorithm) {
       {"typ", Json::FromString(GRPC_JWT_TYPE)},
       {"kid", Json::FromString(key_id)},
   });
-  std::string json_str = JsonDump(json);
+  std::string json_str = grpc_core::JsonDump(json);
   return grpc_base64_encode(json_str.c_str(), json_str.size(), 1, 0);
 }
 
@@ -197,7 +197,8 @@ static char* encoded_jwt_claim(const grpc_auth_json_key* json_key,
     object["sub"] = Json::FromString(json_key->client_email);
   }
 
-  std::string json_str = JsonDump(Json::FromObject(std::move(object)));
+  std::string json_str =
+      grpc_core::JsonDump(Json::FromObject(std::move(object)));
   return grpc_base64_encode(json_str.c_str(), json_str.size(), 1, 0);
 }
